+++ genesis_devtools/builder/dependency.py
@@ -18,7 +18,7 @@ import os
 import shutil
 import typing as tp
 
-from genesis_dev_tools.builder import base
+from genesis_devtools.builder import base
 
 
 class LocalPathDependency(base.AbstractDependency):
