+++ genesis_devtools/packer/ubuntu_24/ubuntu-24.pkr.hcl
