+++ setup.cfg
@@ -5,7 +5,7 @@ description_file =
     README.md
 author = Genesis Corporation
 author_email = anton.kremenetsky@gmail.com
-home_page = https://github.com/infraguys/genesis_dev_tools
+home_page = https://github.com/infraguys/genesis_devtools
 classifier =
     Intended Audience :: Developers
     License :: OSI Approved :: Apache Software License
@@ -21,7 +21,7 @@ classifier =
 
 [files]
 packages =
-    genesis_dev_tools
+    genesis_devtools
 
 
 [package_data]
@@ -31,4 +31,4 @@ ubuntu_24 =
 
 [entry_points]
 console_scripts =
-    genesis = genesis_dev_tools.cmd.cli:main
+    genesis = genesis_devtools.cmd.cli:main
